+++ docs/changes.rst
@@ -1,6 +1,12 @@
 Release history
 ===============
 
+- fix: missing org_id parameters in devices api
+- feat: password parameter in :meth:`wxc_sdk.devices.DevicesApi.create_by_mac_address`
+- feat: new methods in :class:`wxc_sdk.locations.LocationsApi`: list_floors, create_floor, floor_details, update_floor, delete_floor
+- feat: support for virtual extension ranges in result of :meth:`wxc_sdk.telephony.TelephonyApi.test_call_routing`
+- feat: new parameter prefer_e164_format in :meth:`wxc_sdk.person_settings_numbers.NumbersApi.read`
+
 1.14.1
 ------
 - update dependencies to avoid typing-extensions 4.6.0 which breaks Literals in Pydantic models